requirements.txt

atpublic==3.1.1
attrs==23.1.0
billiard==3.6.4.0
black==23.3.0
boto3==1.26.76
botocore==1.29.76
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
configobj==5.0.8
cryptography==40.0.2
dictdiffer==0.9.0
diskcache==5.6.1
distro==1.8.0
dpath==2.1.6
dulwich==0.21.5
filelock==3.12.0
flatten-dict==0.4.2
flufl.lock==7.1.1
frozenlist==1.3.3
fsspec==2023.5.0
tensorflow==2.12.0
funcy==2.0
gitdb==4.0.10
GitPython==3.1.31
grandalf==0.8
h11==0.14.0
hydra-core==1.3.2
idna==3.4
imbalanced-learn==0.11.0
iterative-telemetry==0.0.8
jmespath==1.0.1
kaggle==1.5.16
kombu==5.2.4
markdown-it-py==2.2.0
matplotlib==3.8.0
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nanotime==0.5.2
networkx==3.1
numpy==1.23.2
omegaconf==2.3.0
orjson==3.8.12
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
pycparser==2.21
pydantic==1.10.7
pydot==1.4.2
pygit2==1.12.1
Pygments==2.15.1
pygtrie==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.30.0
rich==13.3.5
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
s3fs==2023.5.0
s3transfer==0.6.1
scmrepo==1.0.3
seaborn==0.13.0
shortuuid==1.0.11
shtab==1.6.1
six==1.16.0
scikit_learn==1.2.2
smmap==5.0.0
sniffio==1.3.0
sqltrie==0.3.1
starlette==0.27.0
tabulate==0.9.0
tomlkit==0.11.8
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.22.0
vine==5.0.0
voluptuous==0.13.1
wcwidth==0.2.6
wrapt==1.14.1
yarl==1.9.2
zc.lockfile==3.0.post1
